HBOnto
The HBOnto is composed of three ontologies. OSMOnto, which contains the various classes of POIs as classified in OSM; ActOnto, which contains the different types of human activities we are interested in;  TimeOnto,  which contains temporal references in which those human activities can occur.  These three ontologies are related by two main relations. The first one connects an action class to the classes of POIs, in which the action can be performed. The second relation connects an action class to the classes of the time of day during which the action is possible.